# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ReportInputTypeId = typeof ReportInputTypeId[keyof typeof ReportInputTypeId];


export const ReportInputTypeId = {
  harassment: 'harassment',
  threats: 'threats',
  blackmail: 'blackmail',
  drugs: 'drugs',
  fraud: 'fraud',
  stalking: 'stalking',
  'something-else': 'something-else',
} as const;

export type ReportInputLanguage = typeof ReportInputLanguage[keyof typeof ReportInputLanguage];


export const ReportInputLanguage = {
  en: 'en',
  bn: 'bn',
} as const;

export interface ReportInput {
  typeId: ReportInputTypeId;
  incidentDate: string;
  /** @minLength 1 */
  division: string;
  /** @minLength 1 */
  district: string;
  /** @minLength 1 */
  policeStation: string;
  /** @minLength 1 */
  village: string;
  location?: string;
  hideIdentity?: boolean;
  /** @minLength 1 */
  summary: string;
  evidence?: string;
  language: ReportInputLanguage;
  consent: boolean;
}

export type ReportLanguage = typeof ReportLanguage[keyof typeof ReportLanguage];


export const ReportLanguage = {
  en: 'en',
  bn: 'bn',
} as const;

export type ReportStatus = typeof ReportStatus[keyof typeof ReportStatus];


export const ReportStatus = {
  Received: 'Received',
  Reviewing: 'Reviewing',
  Action_planned: 'Action planned',
  Closed: 'Closed',
} as const;

export type ReportPriority = typeof ReportPriority[keyof typeof ReportPriority];


export const ReportPriority = {
  Standard: 'Standard',
  Urgent: 'Urgent',
} as const;

export interface Report {
  id: string;
  typeId: string;
  date: string;
  location: string;
  division: string;
  district: string;
  policeStation: string;
  village: string;
  identityHidden: boolean;
  summary: string;
  evidence: string;
  language: ReportLanguage;
  status: ReportStatus;
  priority: ReportPriority;
  actionNote: string;
  nextStep: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ReportTypeCount {
  typeId: string;
  /** @minimum 0 */
  count: number;
}

export interface ReportStats {
  /** @minimum 0 */
  totalReports: number;
  /** @minimum 0 */
  successfulReports: number;
  byType: ReportTypeCount[];
}

export interface PresenceInput {
  /**
     * @minLength 1
     * @maxLength 128
     */
  clientId: string;
}

export interface Presence {
  /** @minimum 0 */
  count: number;
}