# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: reports
    description: Safe report intake operations
  - name: presence
    description: Anonymous live visitor presence
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /reports:
    post:
      operationId: createReport
      tags: [reports]
      summary: Save a safe report
      description: Stores a report submitted through the local safety portal.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Report saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "400":
          description: Invalid report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Report could not be saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reports/stats:
    get:
      operationId: getReportStats
      tags: [reports]
      summary: Get report activity statistics
      description: Returns persisted report totals and counts grouped by report type.
      responses:
        "200":
          description: Report activity statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportStats"
        "500":
          description: Report statistics could not be loaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /presence:
    get:
      operationId: getPresence
      tags: [presence]
      summary: Get live visitor count
      responses:
        "200":
          description: Current anonymous live visitor count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Presence"
    post:
      operationId: updatePresence
      tags: [presence]
      summary: Register an anonymous live visitor heartbeat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PresenceInput"
      responses:
        "200":
          description: Updated anonymous live visitor count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Presence"
        "400":
          description: Invalid presence heartbeat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ReportInput:
      type: object
      properties:
        typeId:
          type: string
          enum: [harassment, threats, blackmail, drugs, fraud, stalking, something-else]
        incidentDate:
          type: string
          format: date
        division:
          type: string
          minLength: 1
        district:
          type: string
          minLength: 1
        policeStation:
          type: string
          minLength: 1
        village:
          type: string
          minLength: 1
        location:
          type: string
        hideIdentity:
          type: boolean
          default: false
        summary:
          type: string
          minLength: 1
        evidence:
          type: string
        language:
          type: string
          enum: [en, bn]
        consent:
          type: boolean
      required:
        - typeId
        - incidentDate
        - division
        - district
        - policeStation
        - village
        - summary
        - language
        - consent
    Report:
      type: object
      properties:
        id:
          type: string
        typeId:
          type: string
        date:
          type: string
          format: date
        location:
          type: string
        division:
          type: string
        district:
          type: string
        policeStation:
          type: string
        village:
          type: string
        identityHidden:
          type: boolean
        summary:
          type: string
        evidence:
          type: string
        language:
          type: string
          enum: [en, bn]
        status:
          type: string
          enum: [Received, Reviewing, Action planned, Closed]
        priority:
          type: string
          enum: [Standard, Urgent]
        actionNote:
          type: string
        nextStep:
          type: string
      required:
        - id
        - typeId
        - date
        - location
        - division
        - district
        - policeStation
        - village
        - identityHidden
        - summary
        - evidence
        - language
        - status
        - priority
        - actionNote
        - nextStep
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    ReportStats:
      type: object
      properties:
        totalReports:
          type: number
          minimum: 0
        successfulReports:
          type: number
          minimum: 0
        byType:
          type: array
          items:
            $ref: "#/components/schemas/ReportTypeCount"
      required:
        - totalReports
        - successfulReports
        - byType
    ReportTypeCount:
      type: object
      properties:
        typeId:
          type: string
        count:
          type: number
          minimum: 0
      required:
        - typeId
        - count
    PresenceInput:
      type: object
      properties:
        clientId:
          type: string
          minLength: 1
          maxLength: 128
      required:
        - clientId
    Presence:
      type: object
      properties:
        count:
          type: number
          minimum: 0
      required:
        - count