# lib/api-zod/src/generated/types/reportStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReportTypeCount } from './reportTypeCount';

export interface ReportStats {
  /** @minimum 0 */
  totalReports: number;
  /** @minimum 0 */
  successfulReports: number;
  byType: ReportTypeCount[];
}